pyproject.toml
[project]
name = "shopwaybot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.21.0",
    "aiohttp>=3.12.15",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "psycopg2>=2.9.10",
    "psycopg2-binary",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
]
